move flask_jwt_extended to separate module
